+++ power_grid_model_c/power_grid_model_c/src/model.cpp
@@ -12,7 +12,7 @@
 
 #include <power_grid_model/auxiliary/dataset.hpp>
 #include <power_grid_model/common/common.hpp>
-#include <power_grid_model/main_model.hpp>
+#include <power_grid_model/main_model_wrapper.hpp>
 
 namespace {
 using namespace power_grid_model;
